new config API and config file format

+++ src/openal_wrapper.h
@@ -76,7 +76,7 @@ AL_API signed char OPENAL_Sample_SetMinMaxDistance(OPENAL_SAMPLE *sptr, float mi
 AL_API signed char OPENAL_SetFrequency(int channel, int freq);
 AL_API signed char OPENAL_SetVolume(int channel, int vol);
 AL_API signed char OPENAL_SetPaused(int channel, signed char paused);
-AL_API void OPENAL_SetSFXMasterVolume(int volume);
+AL_API void OPENAL_SetSFXMasterVolume(float volume);
 AL_API signed char OPENAL_StopSound(int channel);
 AL_API OPENAL_STREAM *OPENAL_Stream_Open(const char *name_or_data, unsigned int mode, int offset, int length);
 AL_API void OPENAL_Stream_Close(OPENAL_STREAM *stream);
